requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asyncpg==0.30.0
attrs==25.3.0
ccxt==4.4.86
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
frozenlist==1.6.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.4
numpy==2.2.6
pandas==2.2.3
propcache==0.3.1
pycares==4.8.0
pycparser==2.22
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
rich==14.0.0
setuptools==80.8.0
six==1.17.0
tabulate==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
